arithmetic associativity, precedence, and bracketing

+++ src/main/kotlin/com/notedgeek/kcmachine/RunCode.kt
@@ -3,4 +3,4 @@ package com.notedgeek.kcmachine
 import com.notedgeek.kcmachine.build.build
 import com.notedgeek.kcmachine.execute.execute
 
-fun runCode(code: String, ramSize: Int) = execute(build(code, mapOf("\$STACK_TOP" to ramSize - 1)), ramSize).toInt()
+fun runCode(code: String, ramSize: Int) = execute(build(code, mapOf("\$STACK_TOP" to ramSize - 1)), ramSize)
